Relocate transparent rendering from events to render Ghost
