+++ src/main/java/com/github/alexthe666/iceandfire/client/render/entity/RenderGhost.java
@@ -7,6 +7,7 @@ import net.minecraft.client.renderer.entity.RenderLiving;
 import net.minecraft.client.renderer.entity.RenderManager;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.MathHelper;
+import org.lwjgl.opengl.GL11;
 
 public class RenderGhost extends RenderLiving<EntityGhost> {
 
@@ -17,7 +18,7 @@ public class RenderGhost extends RenderLiving<EntityGhost> {
 
     public RenderGhost(RenderManager renderManager) {
         super(renderManager, new ModelGhost(0.0F), 0.55F);
-
+        preRenderProfileGhostClean();
     }
 
     public static ResourceLocation getGhostOverlayForType(int ghost) {
@@ -38,6 +39,21 @@ public class RenderGhost extends RenderLiving<EntityGhost> {
         return 0.0F;
     }
 
+    public void preRenderProfileGhostApply(EntityGhost entityIn, float partialTicks) {
+        float alphaForRender = getAlphaForRender(entityIn, partialTicks);
+        GlStateManager.color(1.0F, 1.0F, 1.0F, alphaForRender);
+        GlStateManager.depthMask(false);
+        GlStateManager.enableBlend();
+        GlStateManager.blendFunc(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE_MINUS_SRC_ALPHA);
+        GlStateManager.alphaFunc(GL11.GL_GREATER, 0.001F);
+    }
+
+    public void preRenderProfileGhostClean() {
+        GlStateManager.disableBlend();
+        GlStateManager.alphaFunc(GL11.GL_GREATER, 0.1F);
+        GlStateManager.depthMask(true);
+    }
+
     public float getAlphaForRender(EntityGhost entityIn, float partialTicks) {
         if (entityIn.isDaytimeMode()) {
             return MathHelper.clamp((101 - Math.min(entityIn.getDaytimeCounter(), 100)) / 100F, 0, 1);
@@ -48,6 +64,7 @@ public class RenderGhost extends RenderLiving<EntityGhost> {
     @Override
     public void preRenderCallback(EntityGhost EntityLivingIn, float partialTickTime) {
         this.shadowSize = 0;
+        preRenderProfileGhostApply(EntityLivingIn, partialTickTime);
     }
 
     @Override
